requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
faiss-cpu>=1.9.0
streamlit>=1.41.0
openai>=1.60.0
python-dotenv>=1.0.0
tqdm>=4.67.0
numpy>=1.26.0
